chore(lint): move fl/net/network_detector.h to fl/ to match its fl:: namespace
The file lives under `src/fl/net/` but declares `fl::NetworkDetector`
(not `fl::net::NetworkDetector`), tripping `SubdirNamespaceChecker` which
enforces `src/fl/<subdir>/*.h` -> `namespace fl::<subdir>` and has a
"no suppression" policy.

Moving the file to `src/fl/network_detector.h` resolves the lint
violation without renaming the class. The real RMT5 implementation at
`src/platforms/esp/32/drivers/rmt/rmt_5/network_detector.h` already
declares `fl::NetworkDetector`, so this aligns the facade's namespace
with its class location.

Two consumer includes updated:
- src/fl/channels/driver.cpp.hpp
- src/fl/channels/manager.cpp.hpp

No behavior change; namespace and class unchanged.
